+++ CMakeLists.txt
@@ -105,6 +105,7 @@ add_library(centroids_search_udl SHARED
                             vortex_udls/centroids_search_udl.hpp 
                             vortex_udls/centroids_search_udl.cpp 
                             vortex_udls/serialize_utils.cpp 
+                            vortex_udls/grouped_embeddings_for_search.cpp
                             vortex_udls/api_utils.cpp)
 target_link_libraries(centroids_search_udl PRIVATE ${UDL_COMMON_LIBS} CURL::libcurl)
 target_compile_definitions(centroids_search_udl PRIVATE
@@ -128,6 +129,7 @@ target_compile_definitions(centroids_search_udl PRIVATE
 add_library(clusters_search_udl SHARED 
                                 vortex_udls/clusters_search_udl.hpp 
                                 vortex_udls/clusters_search_udl.cpp 
+                                vortex_udls/grouped_embeddings_for_search.cpp
                                 vortex_udls/serialize_utils.cpp)
 target_link_libraries(clusters_search_udl PRIVATE ${UDL_COMMON_LIBS})
