+++ vortex_udls/centroids_search_udl.cpp
@@ -33,39 +33,14 @@ bool CentroidsSearchOCDPO::ProcessBatchedTasksThread::get_queries_and_emebddings
                                                         std::vector<std::string>& query_list,
                                                         const uint32_t& client_id,
                                                         const uint32_t& query_batch_id) {
-    // If not include_encoder, could directly deserialize the queries and embeddings from the object
-    if (!parent->include_encoder){
-        try{
-            deserialize_embeddings_and_quries_from_bytes(blob->bytes,blob->size,nq,parent->emb_dim,data,query_list);
-        } catch (const std::exception& e) {
-            std::cerr << "Error: failed to deserialize the query embeddings and query texts from the object." << std::endl;
-            dbg_default_error("{}, Failed to deserialize the query embeddings and query texts from the object.", __func__);
-            return false;
-        }
-    } else {
-        // compute the embeddings from the query texts
-        try{
-            // get the query texts from the object
-            nlohmann::json query_texts_json = nlohmann::json::parse(reinterpret_cast<const char*>(blob->bytes), 
-                                                                    reinterpret_cast<const char*>(blob->bytes) + blob->size);
-            query_list = query_texts_json.get<std::vector<std::string>>();
-            nq = query_list.size();
-            data = new float[parent->emb_dim * nq];
-            // compute the embeddings via open AI API call
-            if (!api_utils::get_batch_embeddings(query_list, parent->encoder_name, parent->openai_api_key, parent->emb_dim, data)) {
-                std::cerr << "Error: failed to get the embeddings from the query texts via OpenAI API." << std::endl;
-                dbg_default_error("{}, Failed to get the embeddings from the query texts via OpenAI API.", __func__);
-                delete[] data; 
-                data = nullptr;
-                return false;
-            }
-        } catch (const std::exception& e) {
-            std::cerr << "Error: failed to compute the embeddings from the query texts." << std::endl;
-            dbg_default_error("{}, Failed to compute the embeddings from the query texts.", __func__);
-            return false;
-        }
+    try{
+        deserialize_embeddings_and_quries_from_bytes(blob->bytes,blob->size,nq,parent->emb_dim,data,query_list);
+        return true;
+    } catch (const std::exception& e) {
+        std::cerr << "Error: Centroids search failed to deserialize the query embeddings and query texts from the object." << std::endl;
+        dbg_default_error("{}, Centroids Search Failed to deserialize the query embeddings and query texts from the object.", __func__);
+        return false;
     }
-    return true;
 }
 
 void CentroidsSearchOCDPO::ProcessBatchedTasksThread::combine_common_clusters(const long* I, const int nq, std::map<long, std::vector<int>>& cluster_ids_to_query_ids){
@@ -159,8 +134,6 @@ void CentroidsSearchOCDPO::ProcessBatchedTasksThread::process_task(std::unique_p
     }
     delete[] I;
     delete[] D;
-    if (parent->include_encoder && data) 
-        delete[] data;
     // after using the batched task's blob data, release the memory
     // task_ptr->blob.memory_mode = derecho::cascade::object_memory_mode_t::DEFAULT;
     TimestampLogger::log(LOG_CENTROIDS_EMBEDDINGS_UDL_END,task_ptr->client_id,task_ptr->query_batch_id,parent->my_id);
@@ -252,9 +225,6 @@ void CentroidsSearchOCDPO::set_config(DefaultCascadeContextType* typed_ctxt, con
         if (config.contains("emb_dim")) this->emb_dim = config["emb_dim"].get<int>();
         if (config.contains("top_num_centroids")) this->top_num_centroids = config["top_num_centroids"].get<int>();
         if (config.contains("faiss_search_type")) this->faiss_search_type = config["faiss_search_type"].get<int>();
-        if (config.contains("include_encoder")) this->include_encoder = config["include_encoder"].get<bool>();
-        if (config.contains("encoder_name")) this->encoder_name = config["encoder_name"].get<std::string>();
-        if (config.contains("openai_api_key")) this->openai_api_key = config["openai_api_key"].get<std::string>();
         if (config.contains("emit_key_prefix")) this->emit_key_prefix = config["emit_key_prefix"].get<std::string>();
         if (this->emit_key_prefix.empty() || this->emit_key_prefix.back() != '/') this->emit_key_prefix += '/';
         this->centroids_embs = std::make_unique<GroupedEmbeddingsForSearch>(this->faiss_search_type, this->emb_dim);
