+++ vortex_udls/grouped_embeddings_for_search.hpp
@@ -26,8 +26,34 @@
 namespace derecho{
 namespace cascade{
 
+constexpr size_t CACHE_LINE_SIZE = 64;
+constexpr size_t INITIAL_QUEUE_CAPACITY = 1024;
+
+/*** Class to buffer the receiving queries
+ */
+struct queryQueue{
+    std::vector<std::string> query_list;
+    std::vector<std::string> query_keys;
+    float* query_embs;
+    size_t query_embs_capacity; 
+    std::atomic<size_t> added_query_offset;
+    int emb_dim;
+
+    queryQueue(int emb_dim);
+    ~queryQueue();
+    float* aligned_alloc(size_t size);
+    void resize_queue(size_t new_capacity);
+    bool add_queries(std::vector<std::string>&& queries, const std::string& key, float* embs, int emb_dim, int num_queries);
+    int count_queries();
+    void reset();
+};
+
+
+/*** Wrapper for the ANN search engine
+ * Store group of embeddings for a cluster or for centroids;
+ * Build the index for the embeddings
+ */
 class GroupedEmbeddingsForSearch{
-// Class to store group of embeddings, which could be the embeddings of a cluster or embeddings of all centroids
 public:
      enum class SearchType {
           FaissCpuFlatSearch = 0,
@@ -82,19 +108,7 @@ public:
                                              std::string& cluster_emb_key,
                                              DefaultCascadeContextType* typed_ctxt,
                                              persistent::version_t version,
-                                             bool stable = 1){
-          float* data;
-          // 1. get the object from KV store
-          auto get_query_results = typed_ctxt->get_service_client_ref().get(cluster_emb_key,version, stable);
-          auto& reply = get_query_results.get().begin()->second.get();
-          Blob blob = std::move(const_cast<Blob&>(reply.blob));
-          blob.memory_mode = derecho::cascade::object_memory_mode_t::EMPLACED; // Avoid copy, use bytes from reply.blob, transfer its ownership to GroupedEmbeddingsForSearch.emb_data
-          // 2. get the embeddings from the object
-          data = const_cast<float*>(reinterpret_cast<const float *>(blob.bytes));
-          size_t num_points = blob.size / sizeof(float);
-          retrieved_num_embs += num_points / this->emb_dim;
-          return data;
-     }
+                                             bool stable = 1);
 
      /*** 
      * Helper function to fill_cluster_embs_in_cache()
@@ -105,31 +119,7 @@ public:
                                         std::priority_queue<std::string, std::vector<std::string>, CompareObjKey>& emb_obj_keys,
                                         DefaultCascadeContextType* typed_ctxt,
                                         persistent::version_t version,
-                                        bool stable = 1){
-          float* data;
-          size_t num_obj = emb_obj_keys.size();
-          size_t data_size = 0;
-          Blob blobs[num_obj];
-          size_t i = 0;
-          while (!emb_obj_keys.empty()){
-               std::string emb_obj_key = emb_obj_keys.top();
-               emb_obj_keys.pop();
-               auto get_query_results = typed_ctxt->get_service_client_ref().get(emb_obj_key,version, stable);
-               auto& reply = get_query_results.get().begin()->second.get();
-               blobs[i] = std::move(const_cast<Blob&>(reply.blob));
-               data_size += blobs[i].size / sizeof(float);  
-               i++;
-          }
-          // 2. copy the embeddings from the blobs to the data
-          data = (float*)malloc(data_size * sizeof(float));
-          size_t offset = 0;
-          for (size_t i = 0; i < num_obj; i++) {
-               memcpy(data + offset, blobs[i].bytes, blobs[i].size);
-               offset += blobs[i].size / sizeof(float);
-          }
-          retrieved_num_embs = data_size / this->emb_dim;
-          return data;
-     }
+                                        bool stable = 1);
 
      /***
      * Fill in the embeddings of that cluster by getting the clusters' embeddings from KV store in Cascade
@@ -145,70 +135,11 @@ public:
      * @note we load the stabled version of the cluster embeddings
      ***/
      int retrieve_grouped_embeddings(std::string embs_prefix,
-                                        DefaultCascadeContextType* typed_ctxt){
-          bool stable = 1; 
-          persistent::version_t version = CURRENT_VERSION;
-          // 0. check the keys for this grouped embedding objects stored in cascade
-          //    because of the message size, one cluster might need multiple objects to store its embeddings
-          auto keys_future = typed_ctxt->get_service_client_ref().list_keys(version, stable, embs_prefix);
-          std::vector<std::string> listed_emb_obj_keys = typed_ctxt->get_service_client_ref().wait_list_keys(keys_future);
-          if (listed_emb_obj_keys.empty()) {
-               std::cerr << "Error: prefix [" << embs_prefix <<"] has no embedding object found in the KV store" << std::endl;
-               dbg_default_error("[{}]at {}, Failed to find object prefix {} in the KV store.", gettid(), __func__, embs_prefix);
-               return -1;
-          }
-          std::priority_queue<std::string, std::vector<std::string>, CompareObjKey> emb_obj_keys = filter_exact_matched_keys(listed_emb_obj_keys, embs_prefix);
+                                        DefaultCascadeContextType* typed_ctxt);
 
-          // 1. Get the cluster embeddings from KV store in Cascade
-          float* data;
-          int num_retrieved_embs = 0;
-          if (emb_obj_keys.size() == 1) {
-               std::string emb_obj_key = emb_obj_keys.top();
-               data = single_emb_object_retrieve(num_retrieved_embs, emb_obj_key, typed_ctxt, version, stable);
-          } else {
-               data = multi_emb_object_retrieve(num_retrieved_embs, emb_obj_keys, typed_ctxt, version ,stable);
-          }
-          if (num_retrieved_embs == 0) {
-               std::cerr << "Error: embs_prefix:" << embs_prefix <<" has no embeddings found in the KV store" << std::endl;
-               dbg_default_error("[{}]at {}, There is no embeddings for prefix{} in the KV store.", gettid(), __func__, embs_prefix);
-               return -1;
-          }
-          dbg_default_trace("[{}]: embs_prefix={}, num_emb_objects={} retrieved.", __func__, embs_prefix, num_retrieved_embs);
+     int get_num_embeddings(); 
 
-          // 2. assign the retrieved embeddings
-          this->num_embs = num_retrieved_embs;
-          this->embeddings = data;
-          // this->centroids_embs[cluster_id]= std::make_unique<GroupedEmbeddingsForSearch>(this->emb_dim, retrieved_num_embs, data);
-          // int init_search_res = this->initialize_groupped_embeddings_for_search();
-          return 0;
-     }
-
-     int get_num_embeddings(){
-          return this->num_embs;
-     }   
-
-     int initialize_groupped_embeddings_for_search(){
-          switch (this->search_type) {
-          case SearchType::FaissCpuFlatSearch:
-               initialize_cpu_flat_search();
-               break;
-          case SearchType::FaissGpuFlatSearch:
-               initialize_gpu_flat_search();
-               break;
-          case SearchType::FaissGpuIvfSearch:
-               initialize_gpu_ivf_flat_search();
-               break;
-          case SearchType::HnswlibCpuSearch:
-               initialize_cpu_hnsw_search();
-               break;
-          default:
-               std::cerr << "Error: faiss_search_type not supported" << std::endl;
-               dbg_default_error("Failed to initialize faiss search type, at clusters_search_udl.");
-               return -1;
-          }
-          initialized_index.store(true);
-          return 0;
-     }
+     int initialize_groupped_embeddings_for_search();
 
      /***
       * Search the top K embeddings that are close to one query
@@ -218,63 +149,17 @@ public:
       * @param D: distance array to store the distance of the top_k embeddings
       * @param I: index array to store the index of the top_k embeddings
       */
-     void search(int nq, float* xq, int top_k, float* D, long* I){
-          switch (this->search_type) {
-          case SearchType::FaissCpuFlatSearch:
-               faiss_cpu_flat_search(nq, xq, top_k, D, I);
-               break;
-          case SearchType::FaissGpuFlatSearch:
-               faiss_gpu_flat_search(nq, xq, top_k, D, I);
-               break;
-          case SearchType::FaissGpuIvfSearch:
-               faiss_gpu_ivf_flat_search(nq, xq, top_k, D, I);
-               break;
-          case SearchType::HnswlibCpuSearch:
-               hnsw_cpu_search(nq, xq, top_k, D, I);
-               break;
-          default:
-               std::cerr << "Error: faiss_search_type not supported" << std::endl;
-               dbg_default_error("Failed to search the top K embeddings, at clusters_search_udl.");
-               break;
-          }
-     }
+     void search(int nq, float* xq, int top_k, float* D, long* I);
 
-     void initialize_cpu_hnsw_search() {
-          this->l2_space = std::make_unique<hnswlib::L2Space>(this->emb_dim);
-          this->cpu_hnsw_index = std::make_unique<hnswlib::HierarchicalNSW<float>>(l2_space.get(), this->num_embs, M, EF_CONSTRUCTION);
+     void initialize_cpu_hnsw_search();
 
-          for(size_t i = 0; i < this->num_embs; i++) {
-               this->cpu_hnsw_index->addPoint(this->embeddings + (i * this->emb_dim), i);
-          }
-     }
-     int hnsw_cpu_search(int nq, float* xq, int top_k, float* D, long* I)  {
-          for(size_t i = 0; i < nq; i++) {
-               const float* query_vector = xq + (i * this->emb_dim);
+     int hnsw_cpu_search(int nq, float* xq, int top_k, float* D, long* I);
 
-               auto results = std::move(this->cpu_hnsw_index->searchKnn(query_vector, top_k));
-
-               for(size_t k = 0; k < top_k; k++) {
-                    auto [distance, idx] = results.top();
-                    results.pop();
-
-                    // populate distance vector which is (n, k)
-                    *(D + (i * top_k) + k) = distance;
-
-                    // populate index vector which is  also (n, k)
-                    *(I + (i * top_k) + k) = idx;
-               }
-          }
-
-          return 0;
-     }
      /*** 
       * Initialize the CPU flat search index based on the embeddings.
       * initalize it if use faiss_cpu_flat_search()
      ***/
-     void initialize_cpu_flat_search(){
-          this->cpu_flatl2_index = std::make_unique<faiss::IndexFlatL2>(this->emb_dim); 
-          this->cpu_flatl2_index->add(this->num_embs, this->embeddings); // add vectors to the index
-     }
+     void initialize_cpu_flat_search();
 
      /***
       * FAISS knn flat search on CPU
@@ -285,21 +170,13 @@ public:
       * @param D: distance array to store the distance of the top_k embeddings
       * @param I: index array to store the index of the top_k embeddings
      ***/
-     int faiss_cpu_flat_search(int nq, float* xq, int top_k, float* D, long* I){
-          dbg_default_trace("FAISS CPU flat Search in [GroupedEmbeddingsForSearch] class");
-          this->cpu_flatl2_index->search(nq, xq, top_k, D, I);
-          return 0;
-     }
+     int faiss_cpu_flat_search(int nq, float* xq, int top_k, float* D, long* I);
 
      /*** 
       * Initialize the GPU flat search index based on the embeddings.
       * initalize it if use faiss_gpu_flat_search()
      ***/
-     void initialize_gpu_flat_search(){
-          this->gpu_res = std::make_unique<faiss::gpu::StandardGpuResources>();
-          this->gpu_flatl2_index = std::make_unique<faiss::gpu::GpuIndexFlatL2>(this->gpu_res.get(), this->emb_dim);
-          this->gpu_flatl2_index->add(this->num_embs, this->embeddings); // add vectors to the index
-     }
+     void initialize_gpu_flat_search();
 
      /***
       * FAISS knn flat l2 search on GPU
@@ -310,23 +187,13 @@ public:
       * @param D: distance array to store the distance of the top_k embeddings
       * @param I: index array to store the index of the top_k embeddings
      ***/
-     int faiss_gpu_flat_search(int nq, float* xq, int top_k, float* D, long* I){
-          dbg_default_trace("FAISS GPU flatl2 Search in [GroupedEmbeddingsForSearch] class" );
-          this->gpu_flatl2_index->search(nq, xq, top_k, D, I);
-          return 0;
-     }
+     int faiss_gpu_flat_search(int nq, float* xq, int top_k, float* D, long* I);
 
      /*** 
       * Initialize the GPU ivf search index based on the embeddings.
       * initalize it if use faiss_gpu_ivf_flat_search()
      ***/
-     void initialize_gpu_ivf_flat_search(){
-          int nlist = 100;
-          this->gpu_res = std::make_unique<faiss::gpu::StandardGpuResources>();
-          this->gpu_ivf_flatl2_index = std::make_unique<faiss::gpu::GpuIndexIVFFlat>(this->gpu_res.get(), this->emb_dim, nlist, faiss::METRIC_L2);
-          this->gpu_ivf_flatl2_index->train(this->num_embs, this->embeddings); // train on the embeddings
-          this->gpu_ivf_flatl2_index->add(this->num_embs, this->embeddings); // add vectors to the index
-     }
+     void initialize_gpu_ivf_flat_search();
 
      /***
       * FAISS knn search based on ivf search on GPU
@@ -337,52 +204,13 @@ public:
       * @param D: distance array to store the distance of the top_k embeddings
       * @param I: index array to store the index of the top_k embeddings
      ***/
-     int faiss_gpu_ivf_flat_search(int nq, float* xq, int top_k, float* D, long* I){
-          dbg_default_trace("FAISS GPU ivf Search in [GroupedEmbeddingsForSearch] class" );
-          this->gpu_ivf_flatl2_index->search(nq, xq, top_k, D, I);
-          return 0;
-     }
+     int faiss_gpu_ivf_flat_search(int nq, float* xq, int top_k, float* D, long* I);
 
      /***
       * Reset the GroupedEmbeddingsForSearch object
       * This function is called when the UDL is released
       */
-     void reset(){
-          if (this->embeddings != nullptr) {
-               free(this->embeddings);
-               this->embeddings = nullptr; 
-          }
-          // index cleanup 
-          if (this->search_type == SearchType::FaissCpuFlatSearch && this->cpu_flatl2_index != nullptr) {
-               this->cpu_flatl2_index->reset();
-          } 
-          else if (this->search_type == SearchType::FaissGpuFlatSearch && this->gpu_flatl2_index != nullptr) {
-               cudaError_t sync_err = cudaDeviceSynchronize();
-               if (sync_err == cudaSuccess) {
-                    this->gpu_flatl2_index->reset();  
-                    this->gpu_res.reset();
-                    this->gpu_flatl2_index=nullptr;
-                    this->gpu_res=nullptr;
-                    cudaDeviceReset();  
-               } else {
-                    std::cerr << "Error during cudaDeviceSynchronize: " << cudaGetErrorString(sync_err) << std::endl;
-               }
-          } 
-          else if (this->search_type == SearchType::FaissGpuIvfSearch && this->gpu_ivf_flatl2_index != nullptr) {
-               cudaError_t sync_err = cudaDeviceSynchronize();
-               if (sync_err == cudaSuccess) {
-                    this->gpu_ivf_flatl2_index->reset();  
-                    this->gpu_res.reset();
-                    this->gpu_ivf_flatl2_index=nullptr;
-                    this->gpu_res=nullptr;
-                    cudaDeviceReset();  
-               } else {
-                    std::cerr << "Error during cudaDeviceSynchronize: " << cudaGetErrorString(sync_err) << std::endl;
-               }
-          } else if (this->search_type == SearchType::HnswlibCpuSearch && this -> cpu_flatl2_index != nullptr) {
-               this->cpu_hnsw_index.reset();
-          }
-     }
+     void reset();
 
      ~GroupedEmbeddingsForSearch() {
           // free(embeddings);
